@@ -2,7 +2,7 @@ package com.uawdevstudios.pocketreminder
 
 import android.content.Intent
 import android.os.Bundle
-import android.support.v7.app.AppCompatActivity;
+import androidx.appcompat.app.AppCompatActivity;
 import java.lang.Exception
 
 class SplashScreen : AppCompatActivity() {
